Added to Project : 1. Maps with Shop dan Duelist Location 2. Entities : MainPlayer, Duelist, Player Modified : 1. Game State almost look like Spec
Not Yet :
1. Save Load
2. Duel State
3. Shop State
4. Card Deck State

+++ src/com/terserah/yugi/Entities/Point.java
@@ -8,8 +8,8 @@ package com.terserah.yugi.Entities;
 import java.lang.Math;
 
 public class Point {
-	double x;
-	double y;
+	int  x;
+	int y;
 
 	// membuat titik dengan koordinat (0,0)
 	public Point() {
@@ -17,31 +17,31 @@ public class Point {
 		this.y = 0;
 	}
 	// membuat titik dengan koordinat yang diberikan (x,y)
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 	
 	//setter
-	public void setX (double x){
+	public void setX (int x){
 		this.x = x;
 	}
 	
-	public void setY (double y){
+	public void setY (int y){
 		this.y = y;
 	}
 	
 	//getter
-	public double getX (){
+	public int getX (){
 		return this.x;
 	}
 	
-	public double getY (){
+	public int getY (){
 		return this.y;
 	}
 	
 	// Menggeser Point sejauh x dan y
-	public void move(double x, double y) { 
+	public void move(int x, int y) { 
 		this.x = this.x + x;
 		this.y = this.y + y;
 	}
